style: renamed component to PascalCase
