+++ src/components/CardCareers/index.js
@@ -1,1 +1,1 @@
-export { default } from './cardCareers';
+export { default } from './CardCareers';
